Rename file to match type
