+++ src/lib.rs
@@ -1,11 +1,11 @@
 mod admin_websocket;
-mod app_agent_websocket;
+mod app_websocket;
 mod app_websocket_inner;
 mod error;
 mod signing;
 
 pub use admin_websocket::{AdminWebsocket, AuthorizeSigningCredentialsPayload, EnableAppResponse};
-pub use app_agent_websocket::{AppWebsocket, ZomeCallTarget};
+pub use app_websocket::{AppWebsocket, ZomeCallTarget};
 pub use error::{ConductorApiError, ConductorApiResult};
 pub use holochain_conductor_api::{
     AdminRequest, AdminResponse, AppAuthenticationRequest, AppAuthenticationToken,
